RabbitMQ/pyproject.toml
[project]
name = "rabbit-mq"
version = "0.1.0"
description = "RabbitMQ Playground"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pika>=1.3.2",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.2",
    "ty>=0.0.1a24",
]
